minor fixes

+++ src/maps/maps.zig
@@ -1,8 +1,8 @@
 const grid = @import("../grid.zig");
 
-pub const BATTLE_MAP_WIDTH: u32 = 32;
-pub const BATTLE_MAP_HEIGHT: u32 = 24;
-pub const BATTLE_MAP_LENGTH: u32 = BATTLE_MAP_WIDTH * BATTLE_MAP_HEIGHT;
+pub const BATTLE_MAP_WIDTH: u16 = 32;
+pub const BATTLE_MAP_HEIGHT: u16 = 24;
+pub const BATTLE_MAP_LENGTH: u16 = BATTLE_MAP_WIDTH * BATTLE_MAP_HEIGHT;
 
 pub const PatternMap = struct {
     a: [24]u32,
